+++ consensus/pbft/api/state_api.go
@@ -36,13 +36,13 @@ type StateApiImpl struct {
 	propagateService  pbft.PropagateService
 	eventService      common.EventService
 	parliamentService pbft.ParliamentService
-	repo              *pbft.StateRepository
+	repo              pbft.StateRepository
 }
 
 var ConsensusCreateError = errors.New("Consensus can't be created")
 
 func NewStateApi(publisherID string, propagateService pbft.PropagateService,
-	eventService common.EventService, parliamentService pbft.ParliamentService, repo *pbft.StateRepository) StateApiImpl {
+	eventService common.EventService, parliamentService pbft.ParliamentService, repo pbft.StateRepository) StateApiImpl {
 	return StateApiImpl{
 		publisherID:       publisherID,
 		propagateService:  propagateService,
