+++ consensus/pbft/api/state_api_intenal_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/it-chain/engine/consensus/pbft"
+	"github.com/it-chain/engine/consensus/pbft/infra/mem"
 	"github.com/it-chain/engine/consensus/pbft/test/mock"
 	"github.com/stretchr/testify/assert"
 )
@@ -225,7 +226,7 @@ func setUpApiCondition(isNeedConsensus bool, peerNum int, isNormalBlock bool,
 		return nil
 	}
 
-	repo := pbft.NewStateRepository()
+	repo := mem.NewStateRepository()
 	if isPrepareConditionSatisfied && isNormalBlock {
 
 		savedConsensus := pbft.State{
